+++ core/board_actions/__init__.py
@@ -1,0 +1,3 @@
+from .action import BoardAction, BoardMutationResult
+
+__all__ = ["BoardAction", "BoardMutationResult"]
